% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_length_fit_file.R
\name{length.fit.preparation}
\alias{length.fit.preparation}
\alias{read_length_fit_file}
\title{Read in and process length.fit file}
\usage{
read_length_fit_file(filename, model_name)
}
\arguments{
\item{filname}{The filename of the length.fit file, including the location.}
}
\description{
Hoovers out the observed and predicted composite (i.e. summed over time) catches at length from the length.fit file.
Original work by Yukio and based on a version in NDB's diagnostics4MFCL package.
It should also work with the weight.fit file but hasn't yet been tested.
Does not handle multispecies.
}
