% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFisheries.R
\name{mapFisheries}
\alias{mapFisheries}
\title{Map Fisheries}
\usage{
mapFisheries(file = "labels.tmp")
}
\arguments{
\item{file}{file containing fisheries labels.}
}
\value{
Data frame containing the columns \code{fishery_name}, \code{fishery},
\code{region}, and \code{group}.
}
\description{
Map fisheries using information from the `labels.tmp` file.
}
\examples{
\dontrun{
mapFisheries()
mapFisheries("../../../MFCL/stepwise/09_IdxNoeff/labels.tmp")
}

}
\seealso{
\code{\link{shinyMFCL-package}} gives an overview of the package.
}
\author{
Arni Magnusson.
}
