% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMFCL-package.R
\docType{package}
\name{shinyMFCL-package}
\alias{shinyMFCL-package}
\alias{shinyMFCL}
\title{Tools to Create Shiny Apps for MULTIFAN-CL Assessments}
\description{
Tools that support the creation of Shiny apps to view the data and results
from MULTIFAN-CL stock assessments.
}
\details{
\emph{Prepare data:}
\tabular{ll}{
  \code{\link{mapFisheries}} \tab map fisheries from labels file
}
}
\references{
\url{https://github.com/PacificCommunity/ofp-sam-shinyMFCL}
}
\seealso{
\pkg{shinyMFCL} is used in combination with the packages \pkg{FLR4MFCL} and
\pkg{data.table}.
}
\author{
Finlay Scott and Arni Magnusson.
}
