% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_length_fit_file.R
\name{read_length_fit_file}
\alias{read_length_fit_file}
\title{Read length.fit File}
\usage{
read_length_fit_file(filename, model_name)
}
\arguments{
\item{filename}{name of \verb{length.fit} file.}

\item{model_name}{optional string describing the model.}
}
\description{
Hoover out the observed and predicted composite (i.e. summed over time)
catches at length from the \verb{length.fit} file.
}
\note{
Original work by Yukio and based on a version in NDB's
\verb{diagnostics4MFCL} package.

It should also work with the \verb{weight.fit} file but hasn't yet been
tested.

Does not handle multispecies.
}
