% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_length_fit_file.R
\name{read_length_fit_file}
\alias{read_length_fit_file}
\title{Read length.fit File}
\usage{
read_length_fit_file(filename, model_name = filename)
}
\arguments{
\item{filename}{name of \verb{length.fit} file.}

\item{model_name}{model name (default is same as \code{filename}).}
}
\value{
Data frame containing five columns:
\tabular{ll}{
  \code{fishery} \tab fishery number\cr
  \code{length}  \tab length bin\cr
  \code{obs}     \tab observed frequency\cr
  \code{pred}    \tab predicted frequency\cr
  \code{model}   \tab model name
}
}
\description{
Read observed and predicted composite (summed over time) catches at length
from the \verb{length.fit} file.
}
\author{
Finlay Scott, based on earlier work by Yukio and Nicholas Ducharme-Barth.
}
